Change products URL pattern to be the root URL; add IndexView to display product index page
The `products/` URL pattern in the `eshop/eshop/urls.py` file has been changed to `''` to make it the root URL pattern. In the `products/urls.py` file, a new URL pattern has been added to include the `IndexView` for displaying the product index page. The `IndexView` is implemented in the `views.py` file.