Add URL and view for category list in products app
This diff adds a new URL pattern for the category list view and a new view, CategoryListView, in the products app. The view inherits from Django's ListView and displays a list of all categories in the system. The corresponding template for this view is "products/category\_list.html". The get\_queryset method returns all categories in the system.